Create a separate class for null object
