+++ src/google_cloud_debugger/google_cloud_debugger_lib/dbg_string.h
@@ -15,7 +15,7 @@
 #ifndef DBG_STRING_H_
 #define DBG_STRING_H_
 
-#include "dbg_object.h"
+#include "dbg_reference_object.h"
 
 namespace google_cloud_debugger {
 class EvalCoordinator;
@@ -23,9 +23,9 @@ class EvalCoordinator;
 // This class represents .NET System.String.
 // A strong handle to the underlying string object is stored
 // so we won't lose reference to it.
-class DbgString : public DbgObject {
+class DbgString : public DbgReferenceObject {
  public:
-  DbgString(ICorDebugType *pType) : DbgObject(pType, 0) {}
+  DbgString(ICorDebugType *pType) : DbgReferenceObject(pType, 0) {}
 
   // Creates a strong handle to the object and stores it in string_handle_.
   void Initialize(ICorDebugValue *debug_value, BOOL is_null) override;
@@ -47,9 +47,6 @@ class DbgString : public DbgObject {
   // into string_obj_. Will not do anything if string_obj_set_ is true.
   HRESULT ExtractStringFromReference();
    
-  // Handle to the underlying string object.
-  CComPtr<ICorDebugHandleValue> string_handle_;
-
   // The underlying string object.
   std::string string_obj_;
